backend/requirements.txt
Django==6.0.2
djangorestframework==3.16.1
django-environ==0.13.0
django-cors-headers==4.9.0
drf-spectacular==0.29.0
psycopg[binary]==3.3.3
gunicorn==25.1.0
